Fix bigint/integer mismatch in the points backfill
sum() returns bigint and Postgres will not narrow an argument to find a
function, so level_for_points(sum(points)) failed with "function does not
exist" when running migration 6.

- backfill casts the total inside the subquery
- migration 1 also defines a bigint overload that delegates to the
  integer one, dropped before create so it can replace a hand-made
  version whose parameter has a different name
- recalc_points_total and the badge upvote count cast explicitly too,
  so the intent is visible rather than relying on assignment casts

Audited every other function call in the migrations: no other aggregate
is passed straight into a function.

+++ supabase/migrations/20260807_points_3_badges.sql
@@ -121,7 +121,7 @@ begin
   select count(*) into v_upvotes_given
     from public.votes where user_id = p_user_id and type = 'up';
 
-  select coalesce(sum(coalesce(upvotes, 0)), 0) into v_upvotes_received
+  select coalesce(sum(coalesce(upvotes, 0)), 0)::integer into v_upvotes_received
     from public.experiences where author_id = p_user_id and status = 'active';
 
   -- insertul stă într-un CTE: PL/pgSQL nu poate itera direct peste
